Starting to move switch AST into seperate files
