+++ boa/src/syntax/ast/node/switch.rs
@@ -1,29 +1,51 @@
-Switch(Box<Node>, Box<[(Node, Box<[Node]>)]>, Option<Box<Node>>),
+//! Switch node.
+
+use super::Node;
+use gc::{Finalize, Trace};
+use std::fmt;
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "serde", serde(transparent))]
+#[derive(Clone, Debug, Trace, Finalize, PartialEq)]
+pub struct Switch {
 
- Node::Switch(ref val_e, ref vals, ref default) => {
-    let val = val_e.run(interpreter)?;
-    let mut result = Value::null();
-    let mut matched = false;
-    for tup in vals.iter() {
-        let cond = &tup.0;
-        let block = &tup.1;
-        if val.strict_equals(&cond.run(interpreter)?) {
-            matched = true;
-            let last_expr = block.last().expect("Block has no expressions");
-            for expr in block.iter() {
-                let e_result = expr.run(interpreter)?;
-                if expr == last_expr {
-                    result = e_result;
-                }
-            }
-        }
+}
+
+/// The `switch` statement evaluates an expression, matching the expression's value to a case
+/// clause, and executes statements associated with that case, as well as statements in cases
+/// that follow the matching case.
+///
+/// A `switch` statement first evaluates its expression. It then looks for the first case
+/// clause whose expression evaluates to the same value as the result of the input expression
+/// (using the strict comparison, `===`) and transfers control to that clause, executing the
+/// associated statements. (If multiple cases match the provided value, the first case that
+/// matches is selected, even if the cases are not equal to each other.)
+///
+/// More information:
+///  - [ECMAScript reference][spec]
+///  - [MDN documentation][mdn]
+///
+/// [spec]: https://tc39.es/ecma262/#prod-SwitchStatement
+/// [mdn]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/switch
+impl Switch {
+    /// Implements the display formatting with indentation.
+    pub(super) fn display(&self, f: &mut fmt::Formatter<'_>, indent: usize) -> fmt::Result {
+        
     }
-    if !matched {
-        if let Some(default) = default {
-            result = default.run(interpreter)?;
-        }
+}
+
+impl fmt::Display for Switch {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.display(f, 0)
     }
-    Ok(result)
+}
+
+impl From<Switch> for Node {
+    fn from(switch: Switch) -> Self {
+        Self::Switch(switch)
+    }
 }
